chore(jfr-analytics): wrap view in Tabs
